+++ pulsar/__init__.py
@@ -1,4 +1,4 @@
-'''Concurrent server and message queues'''
+'''Event driven concurrent framework for Python'''
 VERSION = (0, 3, 0)
 
 def get_version():
@@ -23,6 +23,7 @@ CLASSIFIERS  = [
                 'Programming Language :: Python :: 3',
                 'Programming Language :: Python :: 3.1',
                 'Programming Language :: Python :: 3.2',
+                'Programming Language :: Python :: 3.3',
                 'Topic :: Internet',
                 'Topic :: Software Development',
                 'Topic :: Utilities'